+++ Dozent/DownloaderTools.py
@@ -8,23 +8,8 @@ from pySmartDL import SmartDL
 
 
 class DownloaderTools:
-    @classmethod
-    def download_pysmartdl(cls, link: str, verbose=True):
-        """
-        Downloads file from link using PySmartDL
-        :param link: link that needs to be downloaded
-        :param verbose: Show verbose output
-        :return: None
-        """
-        downloader_obj = SmartDL(link, '~/Downloads/', progress_bar=False)
-        downloader_obj.start(blocking=False)
-        while not downloader_obj.isFinished():
-            if verbose:
-                print(DownloaderTools._make_progress_status(downloader_obj, link), end="")
-            time.sleep(.25)
-
     @staticmethod
-    def _make_progress_status(downloader_obj: SmartDL, link, progress_bar_length=20) -> str:
+    def _make_progress_status(downloader_obj: SmartDL, link: str, progress_bar_length: int = 20) -> str:
         """
         Function to make progress bar
         :param downloader_obj: SmartDL object that's currently downloading a file
@@ -45,6 +30,21 @@ class DownloaderTools:
             f"[{progress_percentage}%, {eta} left]"
 
     @classmethod
+    def download_pysmartdl(cls, link: str, verbose: bool = True):
+        """
+        Downloads file from link using PySmartDL
+        :param link: link that needs to be downloaded
+        :param verbose: Show verbose output
+        :return: None
+        """
+        downloader_obj = SmartDL(link, '~/Downloads/', progress_bar=False)
+        downloader_obj.start(blocking=False)
+        while not downloader_obj.isFinished():
+            if verbose:
+                print(cls._make_progress_status(downloader_obj, link), end="")
+            time.sleep(.25)
+
+    @classmethod
     def download_axel(cls, link: str):
         """
         Downloads file from link using axel
@@ -52,7 +52,7 @@ class DownloaderTools:
         :return: None
         """
         os.system(
-            f"axel --verbose --alternate --num-connections={DownloaderTools._connections_count} {link}")
+            f"axel --verbose --alternate --num-connections={cls._connections_count} {link}")
 
     @classmethod
     def download_aria2(cls, link: str):
@@ -61,16 +61,16 @@ class DownloaderTools:
         :param link: link that needs to be downloaded
         :return: None
         """
-        os.system(f"aria2c -x {DownloaderTools._connections_count} {link}")
+        os.system(f"aria2c -x {cls._connections_count} {link}")
 
     _connections_count = 2 * multiprocessing.cpu_count()
 
+    _downloaders = {
+        'pySmartDL': download_pysmartdl.__func__,
+        'Axel': download_axel.__func__,
+        'aria2': download_aria2.__func__
+    }
 
-_downloaders = {
-    'pySmartDL': DownloaderTools.download_pysmartdl,
-    'Axel': DownloaderTools.download_axel,
-    'aria2': DownloaderTools.download_aria2
-}
 
 if __name__ == "__main__":
-    pass
+    RuntimeError('Not implemented')
